update layout and refactor to Comments component
