+++ apps/mobile/src/components/Feed/Posts/FeedPostSocializeSection.tsx
@@ -4,15 +4,14 @@ import { graphql, useFragment } from 'react-relay';
 import { useTogglePostAdmire } from 'src/hooks/useTogglePostAdmire';
 
 import { GalleryBottomSheetModalType } from '~/components/GalleryBottomSheet/GalleryBottomSheetModal';
-import { GalleryTouchableOpacity } from '~/components/GalleryTouchableOpacity';
-import { Typography } from '~/components/Typography';
 import { FeedPostSocializeSectionFragment$key } from '~/generated/FeedPostSocializeSectionFragment.graphql';
 import { FeedPostSocializeSectionQueryFragment$key } from '~/generated/FeedPostSocializeSectionQueryFragment.graphql';
 
 import { CommentsBottomSheet } from '../CommentsBottomSheet/CommentsBottomSheet';
 import { AdmireButton } from '../Socialize/AdmireButton';
+import { Admires } from '../Socialize/Admires';
 import { CommentButton } from '../Socialize/CommentButton';
-import { Interactions } from '../Socialize/Interactions';
+import Comments from '../Socialize/Comments';
 
 type Props = {
   feedPostRef: FeedPostSocializeSectionFragment$key;
@@ -35,7 +34,7 @@ export function FeedPostSocializeSection({ feedPostRef, queryRef }: Props) {
           edges {
             node {
               dbid
-              ...InteractionsAdmiresFragment
+              ...AdmiresFragment
             }
           }
         }
@@ -48,7 +47,7 @@ export function FeedPostSocializeSection({ feedPostRef, queryRef }: Props) {
           }
           edges {
             node {
-              ...InteractionsCommentsFragment
+              ...CommentsFragment
             }
           }
         }
@@ -86,7 +85,6 @@ export function FeedPostSocializeSection({ feedPostRef, queryRef }: Props) {
   }, [post.comments?.edges]);
 
   const totalComments = post.comments?.pageInfo?.total ?? 0;
-  const isEmptyComments = totalComments === 0;
 
   const nonNullAdmires = useMemo(() => {
     const admires = [];
@@ -114,12 +112,10 @@ export function FeedPostSocializeSection({ feedPostRef, queryRef }: Props) {
       <View className="px-3 pb-8 pt-2">
         <View className="flex flex-row justify-between">
           <View className="flex-1 pr-4 pt-1">
-            <Interactions
+            <Admires
               type="Post"
               feedId={post.dbid}
-              commentRefs={nonNullComments}
               admireRefs={nonNullAdmires}
-              totalComments={totalComments}
               totalAdmires={totalAdmires}
               onAdmirePress={toggleAdmire}
               openCommentBottomSheet={handleOpenCommentBottomSheet}
@@ -131,20 +127,11 @@ export function FeedPostSocializeSection({ feedPostRef, queryRef }: Props) {
             <CommentButton openCommentBottomSheet={handleOpenCommentBottomSheet} />
           </View>
         </View>
-        {isEmptyComments && (
-          <GalleryTouchableOpacity
-            onPress={handleOpenCommentBottomSheet}
-            eventElementId={null}
-            eventName={null}
-          >
-            <Typography
-              font={{ family: 'ABCDiatype', weight: 'Regular' }}
-              className="text-sm text-shadow"
-            >
-              Add a comment
-            </Typography>
-          </GalleryTouchableOpacity>
-        )}
+        <Comments
+          commentRefs={nonNullComments}
+          totalComments={totalComments}
+          onCommentPress={handleOpenCommentBottomSheet}
+        />
       </View>
       <CommentsBottomSheet type="Post" feedId={post.dbid} bottomSheetRef={commentsBottomSheetRef} />
     </>
