+++ apps/mobile/src/components/Feed/Socialize/FeedEventSocializeSection.tsx
@@ -4,15 +4,14 @@ import { graphql, useFragment } from 'react-relay';
 import { useToggleAdmire } from 'src/hooks/useToggleAdmire';
 
 import { GalleryBottomSheetModalType } from '~/components/GalleryBottomSheet/GalleryBottomSheetModal';
-import { GalleryTouchableOpacity } from '~/components/GalleryTouchableOpacity';
-import { Typography } from '~/components/Typography';
 import { FeedEventSocializeSectionFragment$key } from '~/generated/FeedEventSocializeSectionFragment.graphql';
 import { FeedEventSocializeSectionQueryFragment$key } from '~/generated/FeedEventSocializeSectionQueryFragment.graphql';
 
 import { CommentsBottomSheet } from '../CommentsBottomSheet/CommentsBottomSheet';
 import { AdmireButton } from './AdmireButton';
+import { Admires } from './Admires';
 import { CommentButton } from './CommentButton';
-import { Interactions } from './Interactions';
+import Comments from './Comments';
 
 type Props = {
   feedEventRef: FeedEventSocializeSectionFragment$key;
@@ -40,7 +39,7 @@ export function FeedEventSocializeSection({ feedEventRef, queryRef, onCommentPre
           edges {
             node {
               dbid
-              ...InteractionsAdmiresFragment
+              ...AdmiresFragment
             }
           }
         }
@@ -53,7 +52,7 @@ export function FeedEventSocializeSection({ feedEventRef, queryRef, onCommentPre
           }
           edges {
             node {
-              ...InteractionsCommentsFragment
+              ...CommentsFragment
             }
           }
         }
@@ -91,7 +90,6 @@ export function FeedEventSocializeSection({ feedEventRef, queryRef, onCommentPre
   }, [event.comments?.edges]);
 
   const totalComments = event.comments?.pageInfo?.total ?? 0;
-  const isEmptyComments = totalComments === 0;
 
   const nonNullAdmires = useMemo(() => {
     const admires = [];
@@ -124,12 +122,10 @@ export function FeedEventSocializeSection({ feedEventRef, queryRef, onCommentPre
       <View className="px-3 pb-8 pt-5">
         <View className="flex flex-row justify-between">
           <View className="flex-1 pr-4 pt-1">
-            <Interactions
+            <Admires
               type="FeedEvent"
               feedId={event.dbid}
-              commentRefs={nonNullComments}
               admireRefs={nonNullAdmires}
-              totalComments={totalComments}
               totalAdmires={totalAdmires}
               onAdmirePress={toggleAdmire}
               openCommentBottomSheet={handleOpenCommentBottomSheet}
@@ -141,20 +137,11 @@ export function FeedEventSocializeSection({ feedEventRef, queryRef, onCommentPre
             <CommentButton openCommentBottomSheet={handleOpenCommentBottomSheet} />
           </View>
         </View>
-        {isEmptyComments && (
-          <GalleryTouchableOpacity
-            onPress={handleOpenCommentBottomSheet}
-            eventElementId={null}
-            eventName={null}
-          >
-            <Typography
-              font={{ family: 'ABCDiatype', weight: 'Regular' }}
-              className="text-sm text-shadow"
-            >
-              Add a comment
-            </Typography>
-          </GalleryTouchableOpacity>
-        )}
+        <Comments
+          commentRefs={nonNullComments}
+          totalComments={totalComments}
+          onCommentPress={handleOpenCommentBottomSheet}
+        />
       </View>
       <CommentsBottomSheet
         type="FeedEvent"
